+++ pyplotter/__init__.py
@@ -1,5 +1,4 @@
 #-*- coding: utf-8 -*-
 from pyplotter import Graph, Plotter
-__version__ = (0, 0, 1)
-version = __version__
-__all__ = ['Graph', 'Plotter', 'version']
+__version__ = (0, 0, 2)
+__all__ = ['Graph', 'Plotter']