+++ CHANGELOG.md
@@ -1,1 +1,19 @@
-v0.0.1, 2013-10-04 -- Initial release.
+## 0.0.2 (2013-10-05)
+
+Features:
+
+  - New command line arguments
+
+Bugfixes:
+
+  - Fix for columnn padding when calling from command line
+
+## 0.0.1 (2013-10-04)
+
+Features:
+
+  - Initial release
+
+Bugfixes:
+
+  - None